@@ -2,7 +2,7 @@ package io.rmiri.skeleton.sample.activity.xml.autoItemsCountXml;
 
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
